Removed load delay slot test and added JSON target back in
The load delay slot didn't actually test what I expected it to since it was
being compiled with the toolchain assembler. I added the JSON target back in to
have a quick way to verify that load delay slots are being generated correctly.
To test this just set the .cargo/config.toml in the demo app and libpsx to use
the system rustc with the JSON target. Since nops won't be inserted with the
system rustc, the animation shouldn't be visible.