Modified libpsx build script to generate BIOS asm wrappers
The input format for each BIOS function is a simplified version of what's in the
nocash specs. This makes it a bit easier to add BIOS functions, but I still have
to add the rust wrapper to bios.rs. I'm also intentionally still tracking bios.s
to make it easier to see what is actually being generated.