Organized libpsx::gpu and started playing with const generic constraints
Const generics are somewhat experimental in rustc right now, so a hodgepodge of
methods are used to relate and constrain the various consts. Within a function
the `constrain` macro works well enough for various (in)equalities between 3
consts. This is what I used in the util mod for various generic array-related
functions. They work as expected and definitely make it easier to not use the
heap, but type inference doesn't always work as expected even though the compile
will error for any return type not allowed by the input types. This lack of
inference is fine within libpsx since this way of constraining consts is
temporary, but it'd be a hassle for using libpsx, so the `draw_*` functions are
wrappers w/o constraint using the minimum number of const parameters required.
The `internal_draw_*` functions contain a parameter for every const expression
"defined" by `draw_*` and constrain these parameters as expected.