Added macros for returning array references from match arms
This is a pattern that shows in `draw_*` since they all take arrays of various
sizes then add the cmd and pass them off to the BIOS. The problem is that match
arm return types have to match so I can't return the array itself. Returning a
reference on the other hand is only allowed if each array is defined outside the
scope of the arm. The macros I added do just that, though I don't like how it
hides mutability behind a macro.