Started working on a centralized way of modifying the GPU and VRAM state
I think the VRAM function I added should be enough to properly demo
framebuffering. The GPU state stuff should be straightforward with the nocash
notes, but trying to add in the typestate stuff seems to be complicating things.
I think it'll probably be worth it though since compile-time checks are
definitely preferred on the console.