Added `alloc` imports to `libpsx::exe` macro
This allows me to just use `Vec` and `Box` without having to lookup the module
path each time. When exe!(no heap) is called, these aren't imported. It seems
that the heap allocations can occasionally be optimized out in simple tests, so
it might make sense to add the imports to `no heap`, but I'd have to see what
happens when an allocation is attempted an fails.