Completely overhauled libpsx to make it more idiomatic rust
This version uses no incomplete features and only relies on const generics in a
few places. The gpu module is somewhat unorganized, but has all the features of
the previous version except some variants of `draw_rect` and `draw_square`. I
like the new design choices in general (IOCtxt, RefCell's in the framebuffer),
although I'd prefer compile-time borrow checker rules instead of the RefCell if
possible.