Added unsafe constructors for all the IO registers
Also added an `()` to each ZST to prevent them from being created w/o `new`. The
idea with `new` is to simplify the structure of the `psx` crate and the `exe`
macro. Basically `exe` has an unsafe block that creates the registers. If the
user wants to shoot themselves in the foot and duplicate the handles to the IO
registers they're free to do so if they use unsafe.