Made demo that decompresses ferris.tim and displays it on the screen
Takes a few seconds, but it actually works. I should clean up the Huffman
compression so I can actually play with the Symbol and Code type sizes. I think
it should be possible to set Code to u16 if I go for a non-optimal encoding
tree. This will help speed up the decompressor since it'll be able to work with
a u32 stream instead of u64. I should also try setting Symbol to u16/u32 to see
if there's any significant improvement in the compression ratio.