Significantly improved decompressor speed by presorting symbol codes
The idea is to sort the arrays of symbols and codes by code (value not length)
then use binary search to find the symbol corresponding to the possible codes in
the decompressor. This improved the demo from decompressing a file (32 KB
zipped/ 128 KB unzipped) in 10 seconds to about 1 second. I also tried
presorting by code length (since it's inversely proportional to the number of
occurences), but the decompressor is still linear in the number symbols so it
gave similar results.

I think hashing may give some speed improvements, but at that point it'd be
difficult to measure and compare with the binary search approach (unless I used
a bigger .TIM in the demo).