Decided to go with arrays in psx::gpu APIs
Modified `unzip` return type to work with this idea. The lazy version returns
the array length as a const to allow its capture and use later on to get subsets
of the uncompressed array. Removing the iterator in psx::gpu::rect_to_vram
shaved about 800 bytes off of the compressed_texture demo.