Started revamping psx::gpu
I realized that individual DrawPort methods become inefficient for a large
number of primitives. I already have one of the two GPU DMA channels in place,
but to make good use of it I need to separate the methods that serialize the
vertex and color data from the methods that send data to the GPU. The overall
design is a step in the right direction, but certainly has room to improve. Also
added basic support for textures.