Made framebuffer attributes configurable
Also removed needless enums in psx::gpu. I realized that with the way I want to
handle primitives, there are going to be a lot of new types so I want to keep
psx::gpu as tidy as possible by removing enums that have little utility (like
just constraining the values of an integer).