Added font example and simplified DMA Transfer
I should come up with better (i.e. non-blocking) ways to make sure that the GPU
cmd buffer doesn't overflow. I think I ran into my first problems with buffer
overflow while debugging the font printer. I chose the quick and easy fix of
polling gpu_stat between primitives, but the ideal way is to organize the
serialized primitives into blocks the size of the buffer and make sure wait
until the buffer is empty before starting.

Also I'm running into some fundamental limits with this fixed-sized packet
approach when trying to format expressions. The issue is that the length of
run-time expressions determines the number of ascii symbols to print. I think
the solution here might be to avoid creating the packets directly and just send
them to the buffer. It's unlikely that those particular primitives will have to
be reused so then the challenge is just how to do this while implementing the
ideal buffer-overflow fix.