Added newline support to font printer
Newline support is kind of a hack since a primitive is still generated for the
newline character. It's just happens to have size-zero, I could probably get rid
of this hack by immediately sending each primitive to the GPU. Then it becomes
a question of UI and how to give Printer access to DrawPort and GpuStat.