Moved from primitives to writing direct to the GPU in font demo
I still need to think about whether I want to use RefCell or &mut for
Framebuffer, TIM and Printer. An increasingly likely alternative is to just say
fuck ownership and create new instances of the IO registers for each of these
structs. I'd have to think carefully about how to avoid possible race conditions
though. At least in the case of a Printer for the panic handler this wouldn't
really be an issue so it might make sense to start adding that to libpsx.