Modified primitive::Buffer to use wrapped pointers
Using ptr instead of dereferencing them fixes the allocation so the struct is
actually stored in the backing array. This also avoids tying every primitive to
the borrows used to allocate them. The wrapped ptr is also only done thru an
Allocator which guarantees that structs don't overlap, but I should probably
review the use of unsafe here again.