Removed Copy bound on Transfer result and made `if_done` method a macro
Removing `Copy` does nothing in practice within the psx crate since it's only
used for Copy types so far, but it should be useful in demos. `if_done` seemed
to be best implemented as a macro since it conditionally consumes the Transfer
which seems impossible to represent with the type system without adding an extra
poison Transfer member.