Started moving away from bitflags in cop0
This allows me to declare cop0::Status `must_use` as a reminder to write the
value back to cop0r12. I changed cop0::Cause as well to keep things uniform, but
since cop0r13 is read-only there's no added benefit. If I were to impl
cop0r3 or cop0r7 (breakpoint stuff) it'd be beneficial to do it this way again,
so I'll probably stick with this method. It'd be nice to change gpu::Stat as
well to get rid of the bitflags dependency, but I'll have to think about how to
do it since gpu::Stat is MMIO.