Removed elf2psexe in favor of a linker script for the psexe header
This will allow better integration with cargo and turns linking into a single
step. It should now be possible to use the custom_test_framework to run tests
(specifying an emulator as the runner ofc). Also added another LLVM patch and
removed rustc's built-in psx linker (change not reflected in patch yet). Also
added `run` and an explicit `build` option to cargo-psx. Added an actual runtime
(`_start`) and some comments in psx crate