Made Framebuffer::swap optionally use DMA and fixed DMA compiler_fence bug
dma::GPU::send_list blocks meaning that stack allocated lists should be fine,
but it was initially only working without LTO. The issue was that with 1 codegen
unit, the compiler seemed to optimize away parts of the InternalDrawEnv
initialization. I think this could be solved by compiler fences around the DMA
transfer, but LLVM doesn't use pseudo-instructions for this and the emitted
instruction is MIPS-II. Forcing a read of the linked list pointer seemed to fix
this. This came at the expense of at least 1 memory load/framebuffer swap :(
but it's progress.