Fixed DMA gpu bug removing need for spurious read_volatile in `GPU::send_list`
This was caused because I forgot `Register::set` was a bitwise set and not an
assignment. I should definitely do some bikeshedding on the HAL trait fn's to
avoid these types of problems. On a related note, now that `send_list` works as
expected it might be time to start thinking about a good trait design for
sending linked lists to the GPU.