Started refactoring libpsx to organize modules' dependencies
The main changes in this commit were moving types defined in the `gpu`,
`interrupt`, `dma` and `timer` modules to `hal` to remove circular dependencies
between `hal` and the others. These types were moved to `hal::*::ty` and then
reexported from the top-level modules. `hal::*::ty` were made `pub(crate)` to
hide the fact that the types are reexported in the documentation. Another big
change was that `DispEnv` and `DrawEnv` were moved from the `gpu` module to
`graphics` and the `Primitive` trait was moved from `graphics` to `hal` and
reexported from the top-level `gpu`.