Added timer::Counter and enabled const_generics
This is completely analogous to dma::Channel. const_generics was enabled to
remove the need for using a hacky trait to associate a const with a type alias.
I'd prefer to remove it if possible since it's an incomplete feature, but things
do look much nicer with the Counter and Channel names as const parameters