Made ColoredVertex private to psx::graphics::primitive
Ideally this would've been an anonymous struct defined in LineG<const N: usize>,
but since that's not allowed this is the next best thing. I still need to decide
on a good way to define LineG<N> to make it clear that the `_pad` on the first
element is actually the GPU command.