Modified fs::DeferredError to use File instead of file descriptors
This should ensure that when the deferred error is read, any file descriptor
used will still be valid. Fixed OpenOptions::open to use the generic error
function since any error in opening a file won't have a valid file descriptor.
Also made DeferredError the default on File::[open, create] since it shouldn't
cost anything if the error is not used.