Change test randomization to use const_random!
This removes the BIOS dependency for generating random numbers and storing the
random seed in the memory card for different runs. const_random! isn't ideal
since it produces a random constant at each invocation site so it won't behave
as expected when used in a loop inside a const fn, but with some slight
workarounds to the original method for generating idx params it works.