Fixed linker issue and improved using features
For some reason, when compiling anything with psx-cargo, it would
generate linker exception because it can't find symbol (putchar).

The issue is in `println!`, and was fixed by replacing the call to
`kernel::printf` with `kernel::std_out_puts`, it is also more efficient
this way, so I changed it as well in `TTY::write_str`.

Also instead if using `cfg!` (which will just be replaced by
`true`/`false`, its better to use `#[cfg]` macro, so that the code will
not be included if not needed.