Add initial BIOS example
This BIOS can't boot up executables yet, but implements enough of the function
vectors to get printing to stdout working. It would be nice to get this working
well enough to boot executables to have a way to test without a commercial BIOS.
Also fussing with the ld script too much may require building with ld.bfd since
lld is very buggy. For now it seems that lld is good enough, but if enough code
needs to be relocated to RAM I should consider switching linkers.