Fix broken --oformat=binary in target JSON
It seems that some rustc changes that landed in the last few days, broke the
--oformat=binary linker arg that the target JSON added. This won't matter when
the builtin target PR gets merged so this commit just adds --oformat=binary to
the cargo invocation through cargo-psx for now. I also updated the README to
suggest installing cargo-psx from this repo to make sure the latest version is
used.