bios: Reduce ctxt switching in exception handler and assign threads time slices
Writing call_handlers in rust means that it has a prologue and epilogue which
duplicates some of the work done by exception_handler. In the case where we
don't switch threads we can ignore s0-s7 and rely on call_handlers to restore
them. I also removed the dedicated exception stack which means we can say the
same about sp and fp. The case where we switch threads, is as costly as before
since we have to load/store s0-s7 in the TCBs. I also reworked the exception
handler to use some of the delay slots and moved the exception handler to RAM.

I also assign each TCB a time-slice so the main thread can take advantage of the
reduced context switching costs. By default it runs 128 times as long as other
threads. The BIOS is still too slow, but I think this may be related to the
Status::ack_all in the IRQ handler.