psx: Make panic thread-safe and reentrant
The example BIOS previously had very random bugs because panic was not
thread-safe. Essentially the loop {} was being ignored because if the exception
handler changed threads. This commit fixes that by ensuring that panic happens
in a critical section. A syscall is used to minimize the chance that an
exception happens between panic's prologue and the critical section call but to
really ensure this panic needs to be a naked function starting with the syscall.
I'm not sure if this is worth it, but it may be a good idea to add a
thread-aware panic handler in the example BIOS.

The panic function was also not reentrant. Normally this should not occur
because the panic function should be simple enough that it doesn't panic, but it
was happening because of a typo in the framebuffer handling introduced when I
added some video mode things. To make it reentrant I just use a static mut flag
since it's already in a critical section and previous iterations won't hold onto
mutable references.