psx: Disable doctests
doctests default to using the mipsel-sony-psx target which requires
no_std/no_main. While we could add that to each individual doctest snippet, when
`cargo test` calls rustdoc it compiles **and links** binaries for each snippet.
This is an issue because `_start` in psx/src/runtime.rs defines
__{c,d}tors_{start,end} and main symbols which end up being undefined for the
doctest binaries. Since doctest binaries don't need any of these (since we only
care if the snippets compile) we could cfg out this code. However `cfg(doc)`
isn't analogous to `cfg(test)` so this isn't straightforward so I'm just
disabling doctests for now. Future options to fix this may including fabricating
cfg features to check if the code being compiled is a doctest snippet (though
the difficulty is differentiating whether `cargo test` is creating a test binary
or a doc binary) or potentially using a different target for doctests (similar
to how the README says documentation should be generated). The latter option
could be the host target or a generic mips target though the latter might not
have `rust-std` available for every nightly (see issue #37) and the former would
mean that rustdoc snippets would need to avoid mips assembly (which might be a
reasonable tradeoff).